widget text and layout minor tweaks color selector added to scripts end in make
This is end of one of the most intence evenings I have ever had ... Shish and me missed anime soc for the first time in years ... it has been really hard :(
Shish is my hero of code ... thank you sir for being such an inspiration. I hope we wont have to work like this again soon.
Greg, your first week was a tour de force ... instantly a member of the crew. You have my admiration and respect.
It's time to sleep ...